some parisng handle

+++ utils/helper_functions.c
@@ -10,11 +10,11 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "cub.h"
+#include "../cub.h"
 
-size_t ft_strlen(const char *str)
+int ft_strlen(const char *str)
 {
-	size_t length = 0;
+	int length = 0;
 
 	while (str[length] != '\0')
 	{
